message list: Rename the merged MessageListWeb to just MessageList.
Now that the functions of the former MessageList and MessageListWeb
are merged into a single explicit component and a `connect` call in
the same single file, give that merged file the name MessageList again.
